Rename st_common.c across stages
