+++ config/splat.us.stcen.yaml
@@ -43,7 +43,7 @@ segments:
       - [0x7B8, .data, update]
       - [0x7E4, .data, collision]
       - [0xBA4, .data, e_red_door]
-      - [0xBBC, .data, 14264]
+      - [0xBBC, .data, st_common]
       - [0xDBC, .data, e_collect]
       - [0x10AC, .data, entity_relic_orb]
       - [0x1110, data]
@@ -72,7 +72,7 @@ segments:
       - [0x11280, c, collision]
       - [0x12A3C, c, create_entity]
       - [0x135B4, c, e_red_door]
-      - [0x14264, c]
+      - [0x14264, c, st_common]
       - [0x15714, c, e_collect]
       - [0x16C90, c, blit_char]
       - [0x16F30, c, entity_relic_orb]
