+++ config/splat.us.stmad.yaml
@@ -83,7 +83,7 @@ segments:
       - [0xEDB8, c, collision]
       - [0x10544, c, create_entity]
       - [0x110A8, c, e_red_door]
-      - [0x11D3C, c]
+      - [0x11D3C, c, st_common]
       - [0x1344C, c, e_collect]
       - [0x149C0, c, entity_relic_orb]
       - [0x15520, c]
