+++ src/st/cen/st_common.c
