+++ src/st/dre/st_common.c
