+++ src/st/mad/st_common.c
