+++ src/st/no3/st_common.c
